wifi(1.3): Add born_name_ to wifi_ap_iface
wifi_ap_iface uses ifname_ for native operations and identification.
This assumes ifname_ is a network interface.

If ap iface is a bridge, then this assumpation is not longer true,
hence add born_name_ for bridge identification, and ifname_ (i.e.
managed interface in bridge) for native operations.

Specially for AP Random MacAddress, it is enabled on both bridge
and its managed interfaces.

CRs-Fixed: 2633678
Change-Id: Ib6ea3ba6e46632f0f068638b166fee66ef78528c